moved IncreaseHelper to structure_manager.hh
