+++ src/structure_managers/adaptor_increase_maxorder.hh
@@ -43,57 +43,7 @@
 
 namespace rascal {
 
-  namespace internal {
-    /**
-     * Specialisation for the static branching to redirect to the
-     * correct number of neighbours of the cluster.
-     */
 
-    template<bool AtMaxOrder>
-    struct IncreaseHelper {
-      template<class Manager_t, class Cluster_t>
-      inline static size_t get_cluster_size(const Manager_t & /*manager*/,
-                                            const Cluster_t & /*cluster*/) {
-        throw std::runtime_error("This branch should never exist.");
-      }
-      template<class Manager_t, class Counters_t>
-      inline static size_t get_offset_impl(const Manager_t & /*manager*/,
-                                           const Counters_t & /*counters*/) {
-        throw std::runtime_error("This branch should never exist.");
-      }
-      template<class Manager_t, class Counters_t>
-      inline static size_t get_cluster_neighbour(const Manager_t & /*manager*/,
-                                                 const Counters_t & /*counters*/,
-                                                 size_t /*index*/) {
-        throw std::runtime_error("This branch should never exist.");
-      }
-    };
-
-    template<>
-    struct IncreaseHelper<false> {
-
-      template<class Manager_t, class Cluster_t>
-      inline static size_t get_cluster_size(const Manager_t & manager,
-                                            const Cluster_t & cluster) {
-        return manager.get_cluster_size(cluster);
-      }
-
-      template<class Manager_t, class Counters_t>
-      inline static size_t get_offset_impl(const Manager_t & manager,
-                                           const Counters_t & counters) {
-        return manager.get_offset_impl(counters);
-      }
-
-      template<class Manager_t, class Counters_t>
-      inline static size_t get_cluster_neighbour(const Manager_t & manager,
-                                                 const Counters_t & counters,
-                                                 size_t index) {
-        return manager.get_cluster_neighbour(counters, index);
-      }
-    };
-
-
-  }  // internal
 
   /**
    * Forward declaration for traits
@@ -268,7 +218,8 @@ namespace rascal {
 
       using IncreaseHelper_t =
         internal::IncreaseHelper<Order == (traits::MaxOrder-1)>;
-      if (Order < traits::MaxOrder-1) {
+
+      if (Order < (traits::MaxOrder-1)) {
         return IncreaseHelper_t::get_cluster_neighbour(this->manager, cluster,
                                                        index);
       } else {
@@ -293,8 +244,7 @@ namespace rascal {
         internal::IncreaseHelper<Order == (traits::MaxOrder-1)>;
 
       if (Order < (traits::MaxOrder-1)) {
-        return IncreaseHelper_t::get_cluster_size(this->manager,
-                                                        cluster);
+        return IncreaseHelper_t::get_cluster_size(this->manager, cluster);
       } else {
         auto access_index = cluster.get_cluster_index(Layer);
         return nb_neigh[access_index];
@@ -428,7 +378,7 @@ namespace rascal {
       //! do nothing, if MaxOrder is not reached, except call the next order
       for (auto next_cluster : cluster) {
 
-        std::cout << "cluster.order " << next_cluster.order() << std::endl;
+        // std::cout << "cluster.order " << next_cluster.order() << std::endl;
         auto & next_cluster_indices
         {std::get<Order>(manager.cluster_indices_container)};
         next_cluster_indices.push_back(next_cluster.get_cluster_indices());
@@ -565,11 +515,11 @@ namespace rascal {
 
 
   /* ---------------------------------------------------------------------- */
-  /* Returns the linear indices of the clusters (whose atom indices
-   * are stored in counters). For example when counters is just the list
-   * of atoms, it returns the index of each atom. If counters is a list of pairs
-   * of indices (i.e. specifying pairs), for each pair of indices i,j it returns
-   * the number entries in the list of pairs before i,j appears.
+  /* Returns the linear indices of the clusters (whose atom indices are stored
+   * in counters). For example when counters is just the list of atoms, it
+   * returns the index of each atom. If counters is a list of pairs of indices
+   * (i.e. specifying pairs), for each pair of indices i,j it returns the number
+   * entries in the list of pairs before i,j appears.
    */
   template<class ManagerImplementation>
   template<size_t Order>
@@ -577,8 +527,8 @@ namespace rascal {
   get_offset_impl(const std::array<size_t, Order> & counters) const {
 
     static_assert(Order < traits::MaxOrder,
-                  "this implementation handles only up to "
-                  "the respective MaxOrder");
+                  "this implementation handles only up to the respective"
+                  " MaxOrder");
     /**
      * Order accessor: 0 - atoms
      *                 1 - pairs
@@ -589,27 +539,28 @@ namespace rascal {
      */
 
     using IncreaseHelper_t =
-      internal::IncreaseHelper_t<Order == (traits::MaxOrder-1)>;
+      internal::IncreaseHelper<Order == (traits::MaxOrder-1)>;
 
-    auto i{GetOffsetImplHelper_t::get_offset_impl(manager, counters)};
-
-      // this->manager.get_offset_impl(counters)};
-
-    if (Order == traits::MaxOrder-1) {
-      /**
-       * Counters as an array to call parent offset multiplet. This can then be
-       * used to access the actual offset for the Order which was built here.
-       */
-      auto j{counters[Order-1]};
-      auto tuple_index{i+j};
-      auto main_offset{this->offsets[tuple_index]};
-      return main_offset;
-    } else {
+    if (Order < (traits::MaxOrder-1)) {
       /**
        * If not accessible at this order, call lower Order offsets from lower
        * order manager or push through to lower levels, if adaptors are stacked.
        */
-      return i;
+      // return IncreaseHelper_t::get_offset_impl(this->manager, counters);
+      // return this->manager.get_offset_impl(counters);
+      return IncreaseHelper_t::get_offset_impl(this->manager, counters);
+    } else {
+      /**
+       * Counters as an array to call parent offset multiplet. This can then be
+       * used to access the actual offset for the Order which was built here.
+       */
+
+      // auto i{IncreaseHelper_t::get_offset_impl(this->manager, counters)};
+      auto i{this->manager.get_offset_impl(counters)};
+      auto j{counters[Order-1]};
+      auto tuple_index{i+j};
+      auto main_offset{this->offsets[tuple_index]};
+      return main_offset;
     }
   }
 }  // rascal
