+++ src/structure_managers/adaptor_neighbour_list.hh
@@ -1025,14 +1025,21 @@ namespace rascal {
      * Order is determined by the ClusterRef building iterator, not by the Order
      * of the built iterator
      */
-    if (Order == 1) {
+
+    using IncreaseHelper_t =
+      internal::IncreaseHelper<Order == (traits::MaxOrder-1)>;
+
+    // auto i = IncreaseHelper_t::get_offset_impl(manager, counters);
+
+    if (Order < (traits::MaxOrder-1)) {
       return this->offsets[counters.front()];
-    } else if (Order == 2) {
+    } else {
       /**
        * Counters as an array to call parent offset multiplet. This can then be
        * used to access the actual offset for the next Order here.
        */
-      auto i{this->manager.get_offset_impl(counters)};
+      auto i{IncreaseHelper_t::get_offset_impl(manager, counters)};
+      //this->manager.get_offset_impl(counters)};
       auto j{counters[Order-1]};
       auto tuple_index{i+j};
       auto main_offset{this->offsets[tuple_index]};
